Refactor: Break down landing page into modular components to fix auth check issue
