+++ components/landing/Footer.tsx
@@ -1,0 +1,55 @@
+'use client'
+import React from "react";
+import Link from "next/link";
+import { CircleIcon } from "lucide-react";
+
+export const Footer = () => {
+  return (
+    <footer className="py-12 bg-gray-900 text-white">
+      <div className="container mx-auto px-4">
+        <div className="grid grid-cols-1 md:grid-cols-4 gap-8 mb-12">
+          <div>
+            <div className="flex items-center gap-2 mb-4">
+              <div className="bg-white rounded-full p-1">
+                <CircleIcon className="h-5 w-5 text-blue-600" />
+              </div>
+              <span className="font-bold text-xl">Resume Builder</span>
+            </div>
+            <p className="text-gray-400 mt-4">
+              AI-powered resume builder helping professionals land their dream jobs.
+            </p>
+          </div>
+          <div>
+            <h4 className="font-bold text-lg mb-4">Product</h4>
+            <ul className="space-y-2">
+              <li><Link href="/features" className="text-gray-400 hover:text-white transition-colors">Features</Link></li>
+              <li><Link href="/pricing" className="text-gray-400 hover:text-white transition-colors">Pricing</Link></li>
+              <li><Link href="/templates" className="text-gray-400 hover:text-white transition-colors">Templates</Link></li>
+            </ul>
+          </div>
+          <div>
+            <h4 className="font-bold text-lg mb-4">Company</h4>
+            <ul className="space-y-2">
+              <li><Link href="/about" className="text-gray-400 hover:text-white transition-colors">About</Link></li>
+              <li><Link href="/careers" className="text-gray-400 hover:text-white transition-colors">Careers</Link></li>
+              <li><Link href="/contact" className="text-gray-400 hover:text-white transition-colors">Contact</Link></li>
+            </ul>
+          </div>
+          <div>
+            <h4 className="font-bold text-lg mb-4">Legal</h4>
+            <ul className="space-y-2">
+              <li><Link href="/privacy" className="text-gray-400 hover:text-white transition-colors">Privacy</Link></li>
+              <li><Link href="/terms" className="text-gray-400 hover:text-white transition-colors">Terms</Link></li>
+              <li><Link href="/cookies" className="text-gray-400 hover:text-white transition-colors">Cookies</Link></li>
+            </ul>
+          </div>
+        </div>
+        <div className="border-t border-gray-800 pt-8 text-center text-sm text-gray-400">
+          &copy; {new Date().getFullYear()} Resume Builder. All rights reserved.
+        </div>
+      </div>
+    </footer>
+  );
+};
+
+export default Footer;
